refactor: move filterProps to ReactUtils and refactor references
