+++ src/cartesian/Bar.tsx
@@ -11,7 +11,7 @@ import { ErrorBar, Props as ErrorBarProps } from './ErrorBar';
 import { Cell } from '../component/Cell';
 import { LabelList } from '../component/LabelList';
 import { uniqueId, mathSign, interpolateNumber } from '../util/DataUtils';
-import { findAllByType } from '../util/ReactUtils';
+import { filterProps, findAllByType } from '../util/ReactUtils';
 import { Global } from '../util/Global';
 import {
   getCateCoordinateOfBar,
@@ -28,7 +28,6 @@ import {
   TooltipType,
   LegendType,
   AnimationTiming,
-  filterProps,
   ChartOffset,
   DataKey,
   TickItem,
