+++ src/numberAxis/Funnel.tsx
@@ -9,7 +9,7 @@ import { Layer } from '../container/Layer';
 import { Trapezoid, Props as TrapezoidProps } from '../shape/Trapezoid';
 import { LabelList } from '../component/LabelList';
 import { Cell, Props as CellProps } from '../component/Cell';
-import { findAllByType } from '../util/ReactUtils';
+import { findAllByType, filterProps } from '../util/ReactUtils';
 import { Global } from '../util/Global';
 import { interpolateNumber } from '../util/DataUtils';
 import { getValueByDataKey } from '../util/ChartUtils';
@@ -19,7 +19,6 @@ import {
   AnimationTiming,
   ChartOffset,
   DataKey,
-  filterProps,
   adaptEventsOfChild,
   PresentationAttributesAdaptChildEvent,
 } from '../util/types';
