+++ src/polar/Pie.tsx
@@ -12,7 +12,7 @@ import { Text } from '../component/Text';
 import { Label } from '../component/Label';
 import { LabelList } from '../component/LabelList';
 import { Cell, Props as CellProps } from '../component/Cell';
-import { findAllByType } from '../util/ReactUtils';
+import { findAllByType, filterProps } from '../util/ReactUtils';
 import { Global } from '../util/Global';
 import { polarToCartesian, getMaxRadius } from '../util/PolarUtils';
 import { isNumber, getPercentValue, mathSign, interpolateNumber, uniqueId } from '../util/DataUtils';
@@ -22,7 +22,6 @@ import {
   LegendType,
   TooltipType,
   AnimationTiming,
-  filterProps,
   Coordinate,
   ChartOffset,
   DataKey,
