+++ src/polar/Radar.tsx
@@ -13,7 +13,8 @@ import { Polygon } from '../shape/Polygon';
 import { Dot, Props as DotProps } from '../shape/Dot';
 import { Layer } from '../container/Layer';
 import { LabelList } from '../component/LabelList';
-import { LegendType, TooltipType, AnimationTiming, filterProps, DataKey } from '../util/types';
+import { LegendType, TooltipType, AnimationTiming, DataKey } from '../util/types';
+import { filterProps } from '../util/ReactUtils';
 import { Props as PolarAngleAxisProps } from './PolarAngleAxis';
 import { Props as PolarRadiusAxisProps } from './PolarRadiusAxis';
