+++ src/shape/Curve.tsx
@@ -20,7 +20,8 @@ import {
 } from 'd3-shape';
 import classNames from 'classnames';
 import _ from 'lodash';
-import { LayoutType, PresentationAttributesWithProps, adaptEventHandlers, filterProps } from '../util/types';
+import { LayoutType, PresentationAttributesWithProps, adaptEventHandlers } from '../util/types';
+import { filterProps } from '../util/ReactUtils';
 import { isNumber } from '../util/DataUtils';
 
 interface CurveFactories {
