+++ src/shape/Rectangle.tsx
@@ -4,7 +4,8 @@
 import React, { PureComponent, SVGProps } from 'react';
 import classNames from 'classnames';
 import Animate from 'react-smooth';
-import { AnimationTiming, filterProps } from '../util/types';
+import { AnimationTiming } from '../util/types';
+import { filterProps } from '../util/ReactUtils';
 
 type RectRadius = [number, number, number, number];
